[chore] change the location of the function
