+++ ratatoskr_tools/datahandle/combine_hists.py
@@ -22,9 +22,35 @@
 ###############################################################################
 import os
 import sys
+import csv
 import pandas as pd
 import numpy as np
 ###############################################################################
+
+def get_latencies(latencies_results_file):
+    """
+    Read the resulting latencies from the csv file.
+
+    Parameters:
+        - results_file: the path to the result file.
+
+    Return:
+        - A list of the filt, packet and network latencies.
+    """
+    latencies = []
+    try:
+        with open(latencies_results_file, newline='') as f:
+            spamreader = csv.reader(f, delimiter=' ', quotechar='|')
+            for row in spamreader:
+                latencies.append(row[1])
+    except Exception:
+        # Add dummy values to latencies, -1.
+        latencies.append(-1)
+        latencies.append(-1)
+        latencies.append(-1)
+
+    return(latencies)
+
 
 
 def combine_VC_hists(directory):
